added changes checkpoint_1
