+++ README.md
@@ -1,2 +1,7 @@
-# 4pockman
+# запуск - run Game.py
+
+# разработана и реализована часть логики игры пакман и интерфейс
+# есть карта-лабиринт, по которой двигается пакман
+# игрок может управлять пакманом с помощью клавиатуры
+
