--- pacman.html
@@ -1,12 +1,0 @@
-<!--[if IE]><meta http-equiv="X-UA-Compatible" content="IE=5,IE=9" ><![endif]-->
-<!DOCTYPE html>
-<html>
-<head>
-<title>pacman.html</title>
-<meta charset="utf-8"/>
-</head>
-<body>
-<div class="mxgraph" style="max-width:100%;border:1px solid transparent;" data-mxgraph="{&quot;highlight&quot;:&quot;#0000ff&quot;,&quot;nav&quot;:true,&quot;resize&quot;:true,&quot;xml&quot;:&quot;&lt;mxfile host=\&quot;app.diagrams.net\&quot; modified=\&quot;2022-03-20T15:57:32.582Z\&quot; agent=\&quot;5.0 (Windows)\&quot; etag=\&quot;byjPCLq1nHierataQEuP\&quot; version=\&quot;17.1.3\&quot; type=\&quot;device\&quot;&gt;&lt;diagram id=\&quot;bSXY4YSZNf_4xe96LLl3\&quot; name=\&quot;Page-1\&quot;&gt;7Vtbc5s6EP41nmkf0jEX3x5jO05Opu1k2syc5rxkZJCNTgTiCMV28usrgbgJjEliDDnNS2Itq9VKu6v9tIKeMXN3lxT4zjdiQ9zT+/auZ8x7uq6Zep//E5SniDLURxFhTZEtmVLCT/QMJVH2Wz8iGwY5RkYIZsjPEy3iedBiORqglGzzbCuC86P6YA0LhJ8WwEXq38hmTkQdx7MQ9CuI1k48sjacRE9cEDPLmQQOsMk2QzIuesaMEsKiX+5uBrFYvHhdon6LPU8TxSj0WK0OjPxze3+/+PdhBpbf8fPVD+PpTErZAPwoJ9zTh5jLm/pCZfYk12H436PQc+oCukZezzjnT/v+jv/lxHC2gn7GiB89MzPPGNyxM4DRWvazuMKQpjL5r7X8H468jAk3wHKBF5P5zJYFVqpSOJev0hwq5hK7Vjystl+DV80dwxXLTL4oW/B6yHrwgAsjxoDRKr2bUwOjjQgqwYW497SiQ2ARClvWwRGh++lzy1psxcbSuhY2onwDRcT71R1V7kpUqRfzDfotA5TNYw1jBT+FGs4iVT+Xaxb29iG0S2bVjL7vYk+EgMULsuSpHaZbftumthHsqGYUBj7YJt63ITzLn1yRekuh58bWN5AyxPHVeQQI5iFkmEp4MI/GnBLOtcIhWlohjmKM6Yp4TKJDTZftBXARFrjyCuINFFKFYzMXCybBw7vOCCY0HNmYT6b9i3NO52mXPMDsk+n5eKEn2mYxVQyQuNZwlyFJjHUJiQsZfeIs8qk+knhPAl4jxn/bFD5qMY+TgY66KYlAQtZ1IjtFdfyHBHYvAHn6XpBno40YUC5+auPaEO3SIUE2SRQQGqeFg3R297G4GzzSNsPoA4h8AJHTApEhcMWm6y0D/10DE2Ua7WWa/RllXw46QqYZjmtkmmFZpjFGXwYN5RqjgwUFVHAZXGz9JWSsgAUTcr6L6rhIHaW4HSTJcOYACqxIx1PXMvLZtyzwIV5dQg9SkN1GZCpU+PuVypYk+oOjuzwA96DYEnFt4tj2NfjD9jdTQdKaWXN/0/pNIWmzZHdTrAI9+1zUnXlriYn1IIIABA605VLy54vQEPN+1MraIrPklDx6dtgr5Nsh9kv8/jLQJ7J9l3k232UbT5nGDaSITx7SZES+EFJU3LxLdOONVFTYimXtNWjA4bMFKxYtLtvzWICsgk+mDmjnKvJF9ygzP4WYb1+bfNG+zPxS3A0JIU7sasYg72rJ9UMsIpqk7JUtsiuCBmpONhRB0SoUBHGXAU8ZNl8wBBUKDxWFx/1qvdQJ5vn5j0iDNDSSBX99tAw6EC31Y+XkcaH/SXEx1Mqx6pHjYvQe4mLYYD0GuHBJAM3iqHpI841VmjJgyXtWn0qLirx1SI5lbygfogxSFkf7AHOnA3NKYhzUPayOm8JyWtlRVbEKRh5MViy+/1dLyx4JmRRbqQZ1kW3jUJgPLOStb4WB52daSvkadpwbKeWHXAYjzHAMMLAMNRMOicES4hsSIHFq4zQa8U7lDsnVH0x7g3lIoWxGPD4JgEL7QRCwLQzYkSxr5i1rDoqWTWrbOcs2ZFjjw7DHqS+phh2bBcOW2VVTc/vxLFsDUX5Y9uUhOyg5WTcUsf52c62ha29hgvupdje/sq9hx99C4hOAAnOesnSnYJum4JJDtnJ6hXvu9vBJNQopeHlJLNRGIUkJ/WBJST34HM339S77/jfwDFuoWZ8WqXf89ZgtwPgFZ5qunDIajOKR3rkorgE5/y+lLqUOdbD2VbXrHax9DcsdoYFS10hNDWrltm6pa2IeEHSkUte4r4xjHih1DSv531zqKrVzFy5MasfFEf17/P78tuAf6i3asa4u1N27i35bVqI9jd++cm9u2d/r3vGdbj8vuI1aSKy9n2t5QaYqaE9cHMsVxyWu2JkDwk/xEcfrLhy6irnfocrxWf6r+Kyn9ZcmhSahX3RCk+/yi6tYmcw3V909PBWuaBaLyWSxaPZYpavAtf1jlVZWGdyfh20EXOLZt45wnEwC1nIJWDebTcB6e4izRgY+3csD2ugAoKybgZOXaBpGproKpfvVyNQYDKr4X4pMeTP9GjhiT7+pNi5+Aw==&lt;/diagram&gt;&lt;/mxfile&gt;&quot;,&quot;toolbar&quot;:&quot;pages zoom layers lightbox&quot;,&quot;page&quot;:0}"></div>
-<script type="text/javascript" src="https://app.diagrams.net/js/viewer-static.min.js"></script>
-</body>
-</html>
