+++ src/Character.py
@@ -1,0 +1,8 @@
+class CharacterInterface:
+    def move(self):
+        pass
+
+    def selfprint(self):
+        pass
+
+
