+++ src/Gameboard.py
@@ -1,0 +1,3 @@
+class GameboardInterface:
+    def selfprint(self):
+        pass
