+++ src/Ghost.py
@@ -1,0 +1,10 @@
+from Character import CharacterInterface
+class Ghost(CharacterInterface):
+    def __init__(self):
+        pass
+
+    def move(self):
+        pass
+
+    def selfprint(self):
+        pass
